🚧 ♻️ WIP refactoring and improving code
