+++ src/components/GeneratePaper/types.ts
